#278: Extract filter utility functions into separate file
